Extracted anonymous test Accession Repository from AccessioningServiceTest to a new class to be reused in other tests
