+++ src/test/java/uk/ac/ebi/ampt2d/accession/AccessioningServiceTest.java
@@ -28,7 +28,6 @@ import java.util.Set;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 
 public class AccessioningServiceTest {
 
@@ -51,25 +50,7 @@ public class AccessioningServiceTest {
             }
         };
 
-        AccessionRepository<String> repository = new AccessionRepository<String>() {
-            private Map<String, String> storage = new HashMap<>();
-
-            @Override
-            public Map<String, String> get(List<String> objects) {
-                Map<String, String> accessions = new HashMap<>();
-                for (String object : objects) {
-                    if (storage.containsKey(object)) {
-                        accessions.put(object, storage.get(object));
-                    }
-                }
-                return accessions;
-            }
-
-            @Override
-            public void add(Map<String, String> accessions) {
-                storage.putAll(accessions);
-            }
-        };
+        AccessionRepository<String> repository = new HashMapTestAccessionRepository();
 
         service = new AccessioningService<>(repository, generator);
     }
