+++ src/test/java/uk/ac/ebi/ampt2d/accession/HashMapTestAccessionRepository.java
@@ -1,0 +1,42 @@
+/*
+ *
+ * Copyright 2017 EMBL - European Bioinformatics Institute
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package uk.ac.ebi.ampt2d.accession;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class HashMapTestAccessionRepository implements AccessionRepository<String> {
+    private Map<String, String> storage = new HashMap<>();
+
+    @Override
+    public Map<String, String> get(List<String> objects) {
+        Map<String, String> accessions = new HashMap<>();
+        for (String object : objects) {
+            if (storage.containsKey(object)) {
+                accessions.put(object, storage.get(object));
+            }
+        }
+        return accessions;
+    }
+
+    @Override
+    public void add(Map<String, String> accessions) {
+        storage.putAll(accessions);
+    }
+}
